refactor: move search state
